cli/command: move prompt utilities to separate package
